adds one click order in reminder mail[#125]
MenuItems from last three orders will be displayed in reminder mail.With
each displayed order, there will be a link to place order. If an user
clicks on that link, the todays order for that employee will be
placed.This today's order will have menu items same as respective order.
This happens without actual signing in.